pyproject.toml
[project]
name = "faq-rag-chatbot"
version = "0.1.0"
description = "RAG-based HR FAQ chatbot using ChromaDB"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.4",
    "langchain>=1.0.7",
    "langchain-chroma>=1.0.0",
    "langchain-openai>=1.0.3",
    "langchain-text-splitters>=1.0.0",
    "numpy>=2.3.4",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
]
